+++ quill/src/detail/misc/RdtscClock.cpp
@@ -82,23 +82,54 @@ RdtscClock::RdtscClock(std::chrono::nanoseconds resync_interval /* = std::chrono
 /***/
 uint64_t RdtscClock::time_since_epoch(uint64_t rdtsc_value) const noexcept
 {
-  // get rtsc current value and compare the diff then add it to base wall time
-  auto diff = static_cast<int64_t>(rdtsc_value - _base_tsc);
+  // get the current index, this is only sef called my the thread that is doing the resync
+  auto const index = _version.load(std::memory_order_relaxed) & (_base.size() - 1);
+
+  // get rdtsc current value and compare the diff then add it to base wall time
+  auto diff = static_cast<int64_t>(rdtsc_value - _base[index].base_tsc);
 
   // we need to sync after we calculated otherwise base_tsc value will be ahead of passed tsc value
   if (diff > _resync_interval_ticks)
   {
     resync(2500);
-    diff = static_cast<int64_t>(rdtsc_value - _base_tsc);
+    diff = static_cast<int64_t>(rdtsc_value - _base[index].base_tsc);
   }
 
-  return static_cast<uint64_t>(_base_time + static_cast<int64_t>(static_cast<double>(diff) * _ns_per_tick));
+  return static_cast<uint64_t>(_base[index].base_time +
+                               static_cast<int64_t>(static_cast<double>(diff) * _ns_per_tick));
+}
+
+/***/
+uint64_t RdtscClock::time_since_epoch_safe(uint64_t rdtsc_value) const noexcept
+{
+  // this function won't resync as it can be called by anyone and only a single thread resyncs
+
+  uint32_t version;
+  uint64_t wall_ts;
+
+  do
+  {
+    version = _version.load(std::memory_order_acquire);
+    auto const index = version & (_base.size() - 1);
+
+    if (QUILL_UNLIKELY((_base[index].base_tsc) == 0 && (_base[index].base_time == 0)))
+    {
+      return 0;
+    }
+
+    // get rdtsc current value and compare the diff then add it to base wall time
+    auto const diff = static_cast<int64_t>(rdtsc_value - _base[index].base_tsc);
+    wall_ts = static_cast<uint64_t>(_base[index].base_time +
+                                    static_cast<int64_t>(static_cast<double>(diff) * _ns_per_tick));
+  } while (version != _version.load(std::memory_order_acquire));
+
+  return wall_ts;
 }
 
 /***/
 bool RdtscClock::resync(uint32_t lag) const noexcept
 {
-  // Sometimes we might get an interrupt and might never resync so we will try again up to max_attempts
+  // Sometimes we might get an interrupt and might never resync, so we will try again up to max_attempts
   constexpr uint8_t max_attempts{4};
 
   for (uint8_t attempt = 0; attempt < max_attempts; ++attempt)
@@ -111,15 +142,19 @@ bool RdtscClock::resync(uint32_t lag) const noexcept
 
     if (QUILL_LIKELY(end - beg <= lag))
     {
-      _base_time = wall_time;
-      _base_tsc = fast_average(beg, end);
+      // update the next index
+      auto const index = (_version.load(std::memory_order_relaxed) + 1) & (_base.size() - 1);
+      _base[index].base_time = wall_time;
+      _base[index].base_tsc = fast_average(beg, end);
+      _version.fetch_add(1, std::memory_order_release);
+
       _resync_interval_ticks = _resync_interval_original;
       return true;
     }
   }
 
-  // we failed to return earlier and we never resynced but we don't really want to keep retrying on each call
-  // to time_since_epoch() so we do non accurate resync we will increace the resynce duration to resync later
+  // we failed to return earlier and we never resynced, but we don't really want to keep retrying on each call
+  // to time_since_epoch() so we do non accurate resync we will increase the resync duration to resync later
   _resync_interval_ticks = _resync_interval_ticks * 2;
   return false;
 }